Move pipeline stuff to new submodule
