+++ src/render/pipeline/shadow.rs
@@ -6,8 +6,9 @@ use nalgebra as na;
 
 use glium::{implement_vertex, uniform, Surface};
 
-use crate::render::instance::UniformsPair;
-use crate::render::{Camera, Context, InstanceParams, RenderLists, Resources};
+use crate::render::{Camera, Resources};
+use crate::render::pipeline::{Context, InstanceParams, RenderLists};
+use crate::render::pipeline::instance::UniformsPair;
 
 #[derive(Debug, Clone)]
 pub struct Config {
